fm-224: move median computation to gas mod
