+++ fendermint/eth/api/src/gas/mod.rs
@@ -1,6 +1,10 @@
 // Copyright 2022-2023 Protocol Labs
 // SPDX-License-Identifier: Apache-2.0, MIT
-use fvm_shared::{econ::TokenAmount, message::Message};
+use fvm_shared::{
+    bigint::{BigInt, Zero},
+    econ::TokenAmount,
+    message::Message,
+};
 
 // Copy of https://github.com/filecoin-project/ref-fvm/blob/fvm%40v3.3.1/fvm/src/gas/outputs.rs
 mod output;
@@ -35,4 +39,40 @@ pub fn effective_gas_premium(msg: &Message, base_fee: &TokenAmount) -> TokenAmou
         return msg.gas_premium.clone();
     }
     available
+}
+
+// finds 55th percntile instead of median to put negative pressure on gas price
+// Rust implementation of:
+// https://github.com/consensus-shipyard/lotus/blob/156f5556b3ecc042764d76308dca357da3adfb4d/node/impl/full/gas.go#L144
+pub fn median_gas_premium(
+    prices: &mut Vec<(TokenAmount, i64)>,
+    block_gas_target: i64,
+) -> TokenAmount {
+    // Sort in descending order based on premium
+    prices.sort_by(|a, b| b.0.cmp(&a.0));
+    let blocks = prices.len() as i64;
+
+    let mut at = block_gas_target * blocks / 2;
+    at += block_gas_target * blocks / (2 * 20);
+
+    let mut prev1 = TokenAmount::zero();
+    let mut prev2 = TokenAmount::zero();
+
+    for (price, limit) in prices.iter() {
+        prev2 = prev1.clone();
+        prev1 = price.clone();
+        at -= limit;
+        if at < 0 {
+            break;
+        }
+    }
+
+    let mut premium = prev1.clone();
+
+    if prev2 != TokenAmount::zero() {
+        premium += &prev2;
+        premium.div_ceil(BigInt::from(2));
+    }
+
+    premium
 }
